+++ Beltalowda.lua
@@ -5,7 +5,7 @@
 Beltalowda = Beltalowda or {}
 Beltalowda.addonName = "Beltalowda"
 Beltalowda.version = 1 --for saved vars, shouldn't be used (version.fix instead)
-Beltalowda.versionString = "0.3.5"
+Beltalowda.versionString = "0.3.6"
 Beltalowda.updateInterval = 50 -- in ms
 Beltalowda.author = "@Kickimanjaro"
 Beltalowda.credits = "Based on RdK Group Tool by @s0rdrak (PC / EU)"